read the undercloud IP from the inventory
set_undercloud_authorized_keys.yml depends on a running undercloud host.
So we can assume we have an undercloud entry in Ansible inventory.

The patch reads the undercloud IP from the inventory, this instead of
the undercloud_ip variable.

This is useful for the cases where the undercloud_ip is dynamical. In
this case, the user has to do the add_host manually at the end of the
pre-run.yml.

Change-Id: Ibe041ba114c65890ef899391318ef88023267208