cert: only on OSP10 and OSP11
The certification needs a valid tempest.conf file to run. The generation
of the tempest file is currently broken on OSP < 10:
  https://github.com/redhat-cip/dci-ansible-agent/issues/6

Because of that, certification fails and we end up with a failed job. It
also worth to mention that OSP8 and OSP9 certification result are not really
relevant anymore. We should instead focus on OSP10 and OSP11.

This commit disable certification keeps the certification only for OSP10
and OSP11, this until at least we fix the tempest.conf generation.

Change-Id: Ib58d695f8893e367686fa241371e54330c330220