import plays/collect_logs.yml
Use the collect-logs from OSP12 everywhere since it works fine with older
version of OS version of OSP. The role makes use of the
inventory_hostname variable, which is the name of the host that run
Ansible. This prevent us to run the role with delegate_to. As a work
around, we overwrite the variable.
We also now prepare the undercloud tarball to be compatible with our
previous usage of the role. see: If9bb85b06422d6ff72e3be156e01a02f9ee31919

Change-Id: Idf4172e182ecb28f35d66c110764dbdf764c4dce