Fix next topic information when not subscribed
If the current topic has a next topic associated but the team doesn't
have access to it, the dci_job module will fail:

The full traceback is:
Traceback (most recent call last):
  File "/tmp/ansible_KVNLtV/ansible_module_dci_topic.py", line 187, in <module>
    main()
  File "/tmp/ansible_KVNLtV/ansible_module_dci_topic.py", line 181, in main
    result = parse_http_response(http_response, dci_topic, context, module)
  File "/tmp/ansible_KVNLtV/ansible_modlib.zip/ansible/module_utils/common.py", line 172, in parse_http_response
  File "/usr/lib/python2.7/site-packages/requests/models.py", line 892, in json
    return complexjson.loads(self.text, **kwargs)
  File "/usr/lib64/python2.7/json/__init__.py", line 338, in loads
    return _default_decoder.decode(s)
  File "/usr/lib64/python2.7/json/decoder.py", line 366, in decode
    obj, end = self.raw_decode(s, idx=_w(s, 0).end())
  File "/usr/lib64/python2.7/json/decoder.py", line 384, in raw_decode
    raise ValueError("No JSON object could be decoded")
ValueError: No JSON object could be decoded

This patch is a temporary fix until this is fixed on dci-ansible
side. That way we can avoid the agent to fail during a legit run.

Change-Id: Idb8f7e561a00dcad5e977b88cb89c7517ce2d5a1