registry: Allow to customize the port
By default, the agent will configure the registry on the port 5000.
In a CI context, when using dci-infra for an AIO infrastructure, this
port will already be used for keystone and docker distribution will
fail to start.
This patch allows to override the registry port configured by the
agent but keep the old default value for backward compatibility.

Change-Id: Id9b0a83584b3027917eeeea05de4e32848738e2f