Fix the teardown hook conditional
The dci_teardown_on_failure variable is used to trigger or not the
teardown hook on failure. This variable is a boolean but we don't
use the bool filter causing the conditional to fail even if the value
of the variable is false because we only check is the variable is not
empty.
This patch removes the 'is defined' mechanism and set a default value
to true (the same that the one in the settings file) and adds the
bool filter to get the false value working.

Change-Id: I5c96ff1c0f524becdc138fdd317bca4c1172daf9