rpm: prepare the dci-openstack-agent rpm
For the transition to dci-openstack-agent, we install a copy of the files
with the dci-openstack-agent path.

The two packages can be installed at the same time.
/etc/dci-openstack-agent is a symlink to /etc/dci-ansible-agent.

This commit also rename dci-ansible-agent.yml to match the expectation
of the dci-openstack-agent-functional test.

To finish the migration, the user will have to:

- install the new dci-openstack-agent rpm
- migrate the ~dci-ansible-agent stuff (e.g: the registry)
- rename the configuration directory

At some point, we will deprecate the dci-ansible-agent:

- move the dci-update.service and dci-update.timer in the new RPM
- use the Obsoletes and Requires to drop the old dci-ansible-agent

Change-Id: Id8edbf6584760f9d0007e1785c6744d64a5d715e