rpm: prepare the dci-openstack-agent rpm
For the transition to dci-openstack-agent, we install a copy of the files
with the dci-openstack-agent path.

The two packages can be installed at the same time. The new own depends
on the old one, this to have access to the dci-update.service file.

This commit also rename dci-ansible-agent.yml to match the expectation
of the dci-openstack-agent-functional test.

To finish the migration, the user will have to:

- install the new dci-openstack-agent rpm
- migrate the ~dci-ansible-agent stuff (e.g: the registry)
- rename the configuration directory

The migration is documentated in I89f498de50eadcba9ce35a7389f0e7f3e19b6e99.

At some point when everybody will be using the new agent, we will deprecate
the dci-ansible-agent:

- move the dci-update.service and dci-update.timer in the new RPM
- use the Obsoletes and Requires to drop the old dci-ansible-agent

Depends-On: I22eff9cb1bb4cc8c6c093fc8959116b55e59ada4

Change-Id: Id8edbf6584760f9d0007e1785c6744d64a5d715e