ffu: only fetch the images for the current topic
With the FFU we jump from OSP10 to OSP13, the upgrade job is attached to
OSP13 BUT it also has the OSP11 and OSP12 components.
We only want to fetch the images for OSP13. So we loop on all the
component and only pick the one with the same topic_id.

Change-Id: I8c8f92a165dfcac1e6c02248845d43151765b307