Update patch set 1
Patch Set 1: Code-Review-1

That's not against the review itself but maybe we need to discuss about components associated to a job that don't belong to the current topic.
For me that should not be possible.

Patch-set: 1
Reviewer: Gerrit User 56 <56@6c1dc8ef-8b94-40e4-bd83-c2359d45ecc0>
Label: Code-Review=-1