Split default and override parameters.
Currently only the settings.yml file was used to add/modify
parameters. But this file is managing by the users and it's not
updated via yum.
As a consequence it's difficult to introduce new parameters without
set a default in the playbook in order to not break older config.
With this change we don't need anymore:
  - '|default("xxxx")' filter.
  - 'is defined' operation.
Also add a sample for the settings.yml file.

Change-Id: I640aa3114c64f823e48a87bc16fe6ddb6d53c144