Update patch set 1
Patch Set 1: Code-Review+1

I like this patch!  You are on fire dimitri \o/

Like Goneri I think it's better to add commented variables needed in settings.yml or in settings.yml.example file.

I don't think a lot of people read the documentation. They probably will start reading the settings.yml file.

An alternative > Can we add a comment with a link to the documentation in settings.yml file if we want to avoid the duplication?

Patch-set: 1
Label: Code-Review=+1