python-docker > 2 breaks us
rhel-7-server-openstack-13-rpms channel comes with a new version of
python-docker that breaks us.

Traceback (most recent call last):
  File "/usr/share/dci-openstack-agent/fetch_images.py", line 141, in <module>
    main()
  File "/usr/share/dci-openstack-agent/fetch_images.py", line 137, in main
    purge_image_from_local_docker(image)
  File "/usr/share/dci-openstack-agent/fetch_images.py", line 91, in purge_image_from_local_docker
    for image in client.images('{origin_registry}/{project}/{name}'.format(**image)):
  File "/usr/lib/python2.7/site-packages/docker/models/resource.py", line 70, in __call__
    .format(self.__class__.__name__))
TypeError: 'ImageCollection' object is not callable. You might be trying to use the old (pre-2.0) API - use docker.APIClient if so.

Also, ensure the package depends on python-docker-py, the package that
we actually use. Finally, we don't the dependency on dci-ansible-agent
anymore.

Change-Id: I946c45e284ff635939a24f00a00bd68b477887e1