Pin ansible-role-openstack-certification to latest version without multi host support
In 9de6b6f7 the role runs sets of tests on specific.

We lack some configuration in the current state of the agent and will
need some time to fix it.
In the meantime, pin to the latest available version without multi host
support to still be able to run certification tests.

Change-Id: Ibe65c037782e64281b26a1ae85bba6e4619addf8